# Conflicts:
#	core/src/main/java/edu/wmi/dpri/przychodnia/server/entity/Person.java

+++ core/src/main/java/edu/wmi/dpri/przychodnia/server/entity/Hire.java
@@ -4,6 +4,7 @@ import org.joda.time.DateTime;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 /**
  * Created by khartv on 21.11.2016.
@@ -34,7 +35,7 @@ public class Hire {
     private Position position;
 
     @Column(name = "salary", precision = 13, scale = 2)
-    private Double salary;
+    private BigDecimal salary;
 
     @ManyToOne(targetEntity = Employee.class)
     @JoinColumn(name = "supervisor")
@@ -83,11 +84,11 @@ public class Hire {
         this.position = position;
     }
 
-    public Double getSalary() {
+    public BigDecimal getSalary() {
         return salary;
     }
 
-    public void setSalary(Double salary) {
+    public void setSalary(BigDecimal salary) {
         this.salary = salary;
     }
